added postgres.py fixtures to be used as shared test code
