+++ packages/simcore-sdk/tests/nodeports/test_nodeports.py
@@ -4,14 +4,14 @@
 import pytest
 
 def test_access_with_key(default_nodeports_configuration): # pylint: disable=W0613, W0621
-    from simcore_sdk.nodeports import PORTS
+    from simcore_sdk.nodeports.nodeports import PORTS
 
     assert PORTS.inputs["in_1"] == PORTS.inputs[0]
     assert PORTS.inputs["in_5"] == PORTS.inputs[1]
     assert PORTS.outputs["out_1"] == PORTS.outputs[0]
 
 def test_port_value_getters(default_nodeports_configuration): # pylint: disable=W0613, W0621
-    from simcore_sdk.nodeports import PORTS
+    from simcore_sdk.nodeports.nodeports import PORTS
 
     assert PORTS.inputs["in_1"].get() == "/home/jovyan/data/outputControllerOut.dat"
     assert PORTS.inputs["in_5"].get() == 666
@@ -23,13 +23,13 @@ def test_port_value_setters(special_nodeports_configuration): # pylint: disable=
     special_config["outputs"].append({
         "key": "out_15",
         "label": "additional data",
-        "description": "here some additional data",
+        "desc": "here some additional data",
         "type": "int",
         "value": "null",
         "timestamp": "2018-05-22T19:34:53.511Z"
     })
     special_nodeports_configuration(special_config)
-    from simcore_sdk.nodeports import PORTS
+    from simcore_sdk.nodeports.nodeports import PORTS
 
     assert PORTS.outputs["out_15"].get() is None
 
@@ -40,8 +40,8 @@ def test_port_value_setters(special_nodeports_configuration): # pylint: disable=
 def test_adding_new_ports(special_nodeports_configuration):
     import helpers
     special_configuration = helpers.get_empty_config() #pylint: disable=E1101
-    config_file = special_nodeports_configuration(special_configuration)
-    from simcore_sdk.nodeports import PORTS
+    engine, session, pipeline_id, node_uuid = special_nodeports_configuration(special_configuration) #pylint: disable=W0612
+    from simcore_sdk.nodeports.nodeports import PORTS
     # check empty configuration
     assert not PORTS.inputs
     assert not PORTS.outputs
@@ -50,45 +50,45 @@ def test_adding_new_ports(special_nodeports_configuration):
     special_configuration["inputs"].append({
         "key": "in_15",
         "label": "additional data",
-        "description": "here some additional data",
+        "desc": "here some additional data",
         "type": "int",
         "value": "15",
         "timestamp": "2018-05-22T19:34:53.511Z"
     })
-    helpers.update_config_file(config_file, special_configuration) #pylint: disable=E1101
+    helpers.update_configuration(session, pipeline_id, node_uuid, special_configuration) #pylint: disable=E1101
 
     assert len(PORTS.inputs) == 1
     assert PORTS.inputs[0].key == "in_15"
     assert PORTS.inputs[0].label == "additional data"
-    assert PORTS.inputs[0].description == "here some additional data"
+    assert PORTS.inputs[0].desc == "here some additional data"
     assert PORTS.inputs[0].type == "int"
     assert PORTS.inputs[0].value == "15"
     assert PORTS.inputs[0].timestamp == "2018-05-22T19:34:53.511Z"
 
-    # replace the configuration now, add an output
+    # # replace the configuration now, add an output
     special_configuration["outputs"].append({
         "key": "out_15",
         "label": "output data",
-        "description": "a cool output",
+        "desc": "a cool output",
         "type": "bool",
         "value": "null",
         "timestamp": "2018-05-22T19:34:53.511Z"
     })
-    helpers.update_config_file(config_file, special_configuration) #pylint: disable=E1101
+    helpers.update_configuration(session, pipeline_id, node_uuid, special_configuration) #pylint: disable=E1101
 
-    # no change on inputs
+    # # no change on inputs
     assert len(PORTS.inputs) == 1
     assert PORTS.inputs[0].key == "in_15"
     assert PORTS.inputs[0].label == "additional data"
-    assert PORTS.inputs[0].description == "here some additional data"
+    assert PORTS.inputs[0].desc == "here some additional data"
     assert PORTS.inputs[0].type == "int"
     assert PORTS.inputs[0].value == "15"
     assert PORTS.inputs[0].timestamp == "2018-05-22T19:34:53.511Z"
-    # new output
+    # # new output
     assert len(PORTS.outputs) == 1
     assert PORTS.outputs[0].key == "out_15"
     assert PORTS.outputs[0].label == "output data"
-    assert PORTS.outputs[0].description == "a cool output"
+    assert PORTS.outputs[0].desc == "a cool output"
     assert PORTS.outputs[0].type == "bool"
     assert PORTS.outputs[0].value == "null"
     assert PORTS.outputs[0].timestamp == "2018-05-22T19:34:53.511Z"
@@ -100,7 +100,7 @@ def test_removing_ports(special_nodeports_configuration):
     special_configuration["inputs"].append({
         "key": "in_15",
         "label": "additional data",
-        "description": "here some additional data",
+        "desc": "here some additional data",
         "type": "int",
         "value": "15",
         "timestamp": "2018-05-22T19:34:53.511Z"
@@ -108,7 +108,7 @@ def test_removing_ports(special_nodeports_configuration):
     special_configuration["inputs"].append({
         "key": "in_17",
         "label": "additional data",
-        "description": "here some additional data",
+        "desc": "here some additional data",
         "type": "int",
         "value": "15",
         "timestamp": "2018-05-22T19:34:53.511Z"
@@ -116,7 +116,7 @@ def test_removing_ports(special_nodeports_configuration):
     special_configuration["outputs"].append({
         "key": "out_15",
         "label": "additional data",
-        "description": "here some additional data",
+        "desc": "here some additional data",
         "type": "int",
         "value": "15",
         "timestamp": "2018-05-22T19:34:53.511Z"
@@ -124,44 +124,44 @@ def test_removing_ports(special_nodeports_configuration):
     special_configuration["outputs"].append({
         "key": "out_17",
         "label": "additional data",
-        "description": "here some additional data",
+        "desc": "here some additional data",
         "type": "int",
         "value": "15",
         "timestamp": "2018-05-22T19:34:53.511Z"
     })
 
-    config_file = special_nodeports_configuration(special_configuration)
-    from simcore_sdk.nodeports import PORTS
+    engine, session, pipeline_id, node_uuid = special_nodeports_configuration(special_configuration) #pylint: disable=W0612
+    from simcore_sdk.nodeports.nodeports import PORTS
     assert len(PORTS.inputs) == 2
     assert len(PORTS.outputs) == 2
     # let's remove the first input
     del special_configuration["inputs"][0]
-    helpers.update_config_file(config_file, special_configuration) #pylint: disable=E1101
+    helpers.update_configuration(session, pipeline_id, node_uuid, special_configuration) #pylint: disable=E1101
     assert len(PORTS.inputs) == 1
     assert len(PORTS.outputs) == 2
 
     assert PORTS.inputs[0].key == "in_17"
     assert PORTS.inputs[0].label == "additional data"
-    assert PORTS.inputs[0].description == "here some additional data"
+    assert PORTS.inputs[0].desc == "here some additional data"
     assert PORTS.inputs[0].type == "int"
     assert PORTS.inputs[0].value == "15"
     assert PORTS.inputs[0].timestamp == "2018-05-22T19:34:53.511Z"
 
     # let's do the same for the second output
     del special_configuration["outputs"][1]
-    helpers.update_config_file(config_file, special_configuration) #pylint: disable=E1101
+    helpers.update_configuration(session, pipeline_id, node_uuid, special_configuration) #pylint: disable=E1101
     assert len(PORTS.inputs) == 1
     assert len(PORTS.outputs) == 1
 
     assert PORTS.outputs[0].key == "out_15"
     assert PORTS.outputs[0].label == "additional data"
-    assert PORTS.outputs[0].description == "here some additional data"
+    assert PORTS.outputs[0].desc == "here some additional data"
     assert PORTS.outputs[0].type == "int"
     assert PORTS.outputs[0].value == "15"
     assert PORTS.outputs[0].timestamp == "2018-05-22T19:34:53.511Z"
 
 def test_changing_inputs_error(default_nodeports_configuration): # pylint: disable=W0613
-    from simcore_sdk.nodeports import PORTS
+    from simcore_sdk.nodeports.nodeports import PORTS
     from simcore_sdk.nodeports.nodeports import DataItemsList
     from simcore_sdk.nodeports import exceptions
 
@@ -173,7 +173,7 @@ def test_changing_inputs_error(default_nodeports_configuration): # pylint: disab
     from simcore_sdk.nodeports._item import DataItem
     new_input = DataItem(key="dummy_1", 
                          label="new label", 
-                         description="new description", 
+                         desc="new description", 
                          type="int", 
                          value="233", 
                          timestamp="2018-06-04T09:46:43:343")
@@ -182,7 +182,7 @@ def test_changing_inputs_error(default_nodeports_configuration): # pylint: disab
     assert "Trying to modify read-only object" in str(excinfo.value)
 
 def test_changing_outputs_error(default_nodeports_configuration): # pylint: disable=W0613
-    from simcore_sdk.nodeports import PORTS
+    from simcore_sdk.nodeports.nodeports import PORTS
     from simcore_sdk.nodeports.nodeports import DataItemsList
     from simcore_sdk.nodeports import exceptions
 
@@ -194,7 +194,7 @@ def test_changing_outputs_error(default_nodeports_configuration): # pylint: disa
     from simcore_sdk.nodeports._item import DataItem
     new_output = DataItem(key="dummy_1", 
                           label="new label", 
-                          description="new description", 
+                          desc="new description", 
                           type="int", 
                           value="233", 
                           timestamp="2018-06-04T09:46:43:343")
